refactor: move starknet constants and helpers to utils
