+++ packages/chain-adapters/src/starknet/index.ts
@@ -1,2 +1,3 @@
 export * from './StarknetChainAdapter'
 export * from './types'
+export * from './utils'
